#11 - Fixed free before writeout of split bug. Fixed input_length is limited by INT_MAX (everything longer was cut away). Fixed console log output was wrong to overflow of float values on large files. README now points towards project with snakemake.

+++ io/raw_input.c
@@ -19,7 +19,7 @@
 KSEQ_INIT(gzFile, gzread)
 
 
-void get_meta(const char *filename, size_t *max_sequence_length, size_t *sequence_count, float *char_count) { // TODO: Count number of symbols for statistic.
+void get_meta(const char *filename, size_t *max_sequence_length, size_t *sequence_count, size_t *char_count) { // TODO: Count number of symbols for statistic.
     // Find max_sequence_length and sequence_count.
     size_t length = 0, count = 0, sum_length = 0;
     kseq_t *seq;
